Crate shared_buffer

Structs

OwnedBuffer

A cheaply cloneable owned buffer that may be backed by a memory-mapped file or Bytes in memory.

OwnedIntoIter

Enums

MmapError

Errors that may occur when using one of the mmap-based implementations of TryFrom.